feat:retirado qualquer metodo relacionado ao menu da classe PIApplication e criado a classe UsuarioMenu
